i have found a (not-so-hacky, but also not-so-clean) solution for being able to compute optimal action sets for various fair recourse svms (still have to get gupta's working)
in short: most of the code (over 180 different parts) assumed that factual_instance is a dictionary containing all endegenous nodes, and nothing else. this is a valid assumption when the model is trained on endegous nodes

for fair models, we have various other node combinations that get input to the classifier, incl'd, in the case of cw-fair, exogenous nodes

this meant that we need to fundamentally change the main data structure being passed around the code to contain more information; while this sounds easy, such a change would involve A LOT of logic changes here and there (e.g., while before, the counterfactual_template would only copy the keys/names of the endogenous variables, now it would have to copy both keys and values of exogenous variables, e.g.. no interventions on exogenous variables allowed, etc.). this solution would need at least 1 week to code up, and at least another to test

the other solution, currently being worked on, is to create a class called Instance which holds both endo/exo variables, and with some clever class methods and getter/setters, one can leave most of the code untouched after initially creating the Instance object (instead of dictionary as before)